chore: Update version of docker-compose used
The deployment method used for installing docker-compose was relying on the old (pre v2)
version. This updates to use the v2 plugin for the Docker engine so that it is easier to
stay up to date with updated releases.
