+++ src/bilder/components/docker/files/docker-compose.service
@@ -18,15 +18,15 @@ RemainAfterExit=true
 WorkingDirectory=/etc/docker/compose/
 EnvironmentFile=-/etc/default/docker-compose
 
-ExecStop=/usr/local/bin/docker-compose down
-ExecStartPre=/usr/local/bin/docker-compose pull --quiet --parallel
+ExecStop=/usr/bin/docker compose down
+ExecStartPre=/usr/bin/docker compose pull --quiet --parallel
 
-ExecStart=/usr/local/bin/docker-compose up -d
+ExecStart=/usr/bin/docker compose up -d
 
-ExecStop=/usr/local/bin/docker-compose down
+ExecStop=/usr/bin/docker compose down
 
-ExecReload=/usr/local/bin/docker-compose pull --quiet --parallel
-ExecReload=/usr/local/bin/docker-compose up -d
+ExecReload=/usr/bin/docker compose pull --quiet --parallel
+ExecReload=/usr/bin/docker compose up -d
 
 [Install]
 WantedBy=multi-user.target
