+++ src/bridge/lib/versions.py
@@ -1,8 +1,7 @@
-AIRBYTE_VERSION = "0.39.41-alpha"
+AIRBYTE_VERSION = "0.39.42-alpha"
 CONCOURSE_VERSION = "7.8.0"
 CONSUL_TEMPLATE_VERSION = "0.29.0"
 CONSUL_VERSION = "1.12.2"
-DAGSTER_VERSION = "0.15.3"
-DOCKER_COMPOSE_VERSION = "1.29.2"
+DAGSTER_VERSION = "1.0.2"
 REDASH_VERSION = "9d273e4"
 VAULT_VERSION = "1.10.4"
